pyproject.toml
[tool.poetry]
name = "processamento-csv"
version = "0.1.0"
description = ""
authors = ["Dérick Eduardo <dreduardo2222@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
ipython = "^8.29.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
